% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/two_variable_error_ratio_specified_exact_statistics.R
\name{two_variable_exact_CI}
\alias{two_variable_exact_CI}
\title{calculate slope, exact CI, etc. for two variable relationship given lambda}
\usage{
two_variable_exact_CI(
  correlation,
  sdOFy_divided_by_sdOFx,
  square_root_of_lambda,
  sample_size,
  slope_hypothesis = NA,
  CI_coefficient = 0.95,
  sql_round_to_zero_threshold = 1e-10,
  lostDoF = 2,
  print_citation_info = TRUE,
  checkArg = TRUE,
  debugArg = FALSE
)
}
\arguments{
\item{correlation}{\code{numeric} correlation of x-axis and y-axis variable.}

\item{sdOFy_divided_by_sdOFx}{\code{numeric}sd(y-axis variable)/sd(x-axis variable).}

\item{square_root_of_lambda}{\code{numeric} sqrt(lambda), where lambda
is the ratio of the error in the y-axis variable to that in the x-axis
variable.}

\item{sample_size}{\code{numeric} number of observations per variable}

\item{slope_hypothesis}{\code{numeric} slope (if any) to test as null hypothesis.}

\item{CI_coefficient}{\code{numeric} defaults to 0.95, \% of time not to reject the
null hypothesis assuming it is true.}

\item{sql_round_to_zero_threshold}{\code{numeric} if the square root of lambda
argument \code{square_root_of_lambda} is less than this threshold it is rounded
to zero as appropriate and so the answer should be identical to as if
zero was given as \code{square_root_of_lambda}}

\item{lostDoF}{\code{numeric} defaults to 2, I have never tried changing
this value, but you can change it IF you have a compelling reason.
Most significance tests require subtracting 2
degrees of freedom from the sample size.}

\item{print_citation_info}{\code{logical} defaults to \code{TRUE}, which means it
prints citation information for this function. Suggested citations is/are the
works that contribute to at least some output or calculation of the
function in some way and deserve citation.}

\item{checkArg}{\code{logical} defaults to \code{TRUE}, which checks arguments for
some basic formatting mistakes.}

\item{debugArg}{\code{logical} defaults to \code{FALSE}, \code{TRUE} prints internal
functional information (if any) to terminal.}
}
\value{
\code{list} NOTE: all polar angles are with respect to the x-axis line
from the origin in the positive direction. Thus, going clockwise from
the x-axis towards the line y=x gives an angle of -135 degrees and going
counterclockwise from the x-axis gives an angle of 45 degrees. Confidence
intervals (CIs) always stretch from the clockwise line
towards the counterclockwise line.
\describe{
\item{slope_cartesian}{slope of best fit line}
\item{slope_polar}{All "polar" angles measure angle between line
and x-axis from origin going in positive direction.
(e.g. line y=x could be 45 degrees in quadrant 1 or -135 in quadrant 3).}
\item{clockwise_CI_cartesian}{ clockwise confidence interval (see "NOTE")
expressed in cartesian terms.}
\item{clockwise_CI_polar}{see "NOTE"}
\item{counterclockwise_CI_cartesian}{see "NOTE"}
\item{counterclockwise_CI_polar}{see "NOTE"}
\item{CI_polar_width}{width (in degrees) of range of slopes
in the CI (confidence interval).}
\item{x_reliability_ratio}{MLE estimate (so divided by n instead of (n-1))
of proportion of x-axis variable's observed variance that is part of
the related (true) x-values and not the error. Remember
v(x_o)=v(x_t)+v(x_e), so this gives v(x_t)/v(x_o).}
\item{slope_test_p_value}{p-value of tested slope hypothesis (if given)}
\item{calculationTerms}{May contain code's intermediate calculation terms
and such (useful for debugging)}
\item{givenArguments}{\code{list} of function argument values used in call.}
}
}
\description{
calculate slope, exact CI, etc. for two variable relationship given lambda
}
\examples{
set.seed(0); n<-10; y1<-rnorm(n); x1<-rnorm(n); y2<--y1
c1a<-two_variable_exact_CI(correlation=cor(x1,y2),
 sdOFy_divided_by_sdOFx=sd(y2)/sd(x1),
 square_root_of_lambda=1,
 sample_size=n,
 CI_coefficient=0.80,
 print_citation_info=FALSE,
 slope_hypothesis=-1.45)
c1a$slope_cartesian
}
