% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/two_variable_error_ratio_specified_exact_statistics.R
\name{validate_two_variable_exact_CI_function_arguments}
\alias{validate_two_variable_exact_CI_function_arguments}
\title{Validate Inputs for EIV Hypothesis Testing}
\usage{
validate_two_variable_exact_CI_function_arguments(
  correlation = NA,
  validate_correlation = FALSE,
  sdOFy_divided_by_sdOFx = NA,
  validate_sdOFy_divided_by_sdOFx = FALSE,
  square_root_of_lambda = NA,
  validate_square_root_of_lambda = FALSE,
  sample_size = NA,
  validate_sample_size = FALSE,
  slope_hypothesis = NA,
  validate_slope_hypothesis = FALSE,
  CI_coefficient = NA,
  validate_CI_coefficient = FALSE,
  sql_round_to_zero_threshold = NA,
  validate_sql_round_to_zero_threshold = FALSE,
  lostDoF = NA,
  validate_lostDoF = FALSE,
  print_citation_info = NA,
  validate_print_citation_info = FALSE,
  checkArg = NA,
  validate_checkArg = FALSE,
  debugArg = NA,
  validate_debugArg = FALSE,
  y = NA,
  validate_y = FALSE,
  x = NA,
  validate_x = FALSE
)
}
\arguments{
\item{correlation}{\code{numeric} correlation of x-axis and y-axis variable.}

\item{validate_correlation}{\code{logical} \code{TRUE} if argument should be validated.}

\item{sdOFy_divided_by_sdOFx}{\code{numeric}sd(y-axis variable)/sd(x-axis variable).}

\item{validate_sdOFy_divided_by_sdOFx}{\code{logical} \code{TRUE} if argument should be validated.}

\item{square_root_of_lambda}{\code{numeric} sqrt(lambda), where lambda
is the ratio of the error in the y-axis variable to that in the x-axis
variable.}

\item{validate_square_root_of_lambda}{\code{logical} \code{TRUE} if argument should be validated.}

\item{sample_size}{\code{numeric} number of observations per variable}

\item{validate_sample_size}{\code{logical} \code{TRUE} if argument should be validated.}

\item{slope_hypothesis}{\code{numeric} slope (if any) to test as null hypothesis.}

\item{validate_slope_hypothesis}{\code{logical} \code{TRUE} if argument should be validated.}

\item{CI_coefficient}{\code{numeric} defaults to 0.95, \% of time not to reject the
null hypothesis assuming it is true.}

\item{validate_CI_coefficient}{\code{logical} \code{TRUE} if argument should be validated.}

\item{sql_round_to_zero_threshold}{\code{numeric} if the square root of lambda
argument \code{square_root_of_lambda} is less than this threshold it is rounded
to zero as appropriate and so the answer should be identical to as if
zero was given as \code{square_root_of_lambda}}

\item{validate_sql_round_to_zero_threshold}{\code{logical} \code{TRUE}
if argument should be validated.}

\item{lostDoF}{\code{numeric} defaults to 2, I have never tried changing
this value, but you can change it IF you have a compelling reason.
Most significance tests require subtracting 2
degrees of freedom from the sample size.}

\item{validate_lostDoF}{\code{logical} \code{TRUE} if argument should be validated.}

\item{print_citation_info}{\code{logical} defaults to \code{TRUE}, which means it
prints citation information for this function. Suggested citations is/are the
works that contribute to at least some output or calculation of the
function in some way and deserve citation.}

\item{validate_print_citation_info}{\code{logical} \code{TRUE} if argument should be validated.}

\item{checkArg}{\code{logical} defaults to \code{TRUE}, which checks arguments for
some basic formatting mistakes.}

\item{validate_checkArg}{\code{logical} \code{TRUE} if argument should be validated.}

\item{debugArg}{\code{logical} defaults to \code{FALSE}, \code{TRUE} prints internal
functional information (if any) to terminal.}

\item{validate_debugArg}{\code{logical} \code{TRUE} if argument should be validated.}

\item{y}{\code{numeric} vector of y-axis observations}

\item{validate_y}{\code{logical} \code{TRUE} if argument should be validated.}

\item{x}{\code{numeric} vector of x-axis observations}

\item{validate_x}{\code{logical} \code{TRUE} if argument should be validated.}
}
\value{
\code{invisible(TRUE)} if all checks pass; throws an
informative error via stop() if any validation fails.
}
\description{
note that these variable descriptions are for use in other roxygen comments
and largely do not reflect the defaults for this validation function
which are always either \code{NA} or \code{FALSE}
}
\details{
this function only checks some basic issues like that the \code{correlation}
argument is between -1 and 1, etc.
}
\examples{
\dontrun{
try(validate_two_variable_exact_CI_function_arguments(correlation=2,
validate_correlation=TRUE))
validate_two_variable_exact_CI_function_arguments(correlation=0.2,
validate_correlation=TRUE)
validate_two_variable_exact_CI_function_arguments(correlation=2,
validate_correlation=FALSE)
validate_two_variable_exact_CI_function_arguments(correlation=0.2,
validate_correlation=FALSE)
}
}
\keyword{internal}
